refactor: create module structure & combine nav components in namespace
