+++ src/modules/Navigation/Navigation.js
@@ -4,8 +4,8 @@ import React from 'react'
 
 import cx from 'classnames'
 
-import { NavigationLink } from './Navigation/NavigationLink'
-import { groupOrder } from '../constants/groupOrder'
+import { NavigationLink } from './NavigationLink'
+import { groupOrder } from '../../constants/groupOrder'
 
 const NavigationComponent = ({ disabled }) => {
     return (
