+++ src/modules/Sidebar/sectionPropType.js
