+++ src/modules/icons/Folder.module.css
