+++ src/modules/icons/FolderClosed.js
