+++ src/modules/icons/FolderOpen.js
