+++ src/modules/icons/Plus.js
