+++ src/pages/All.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import { CardMenu } from '../components/CardMenu'
-import { Container } from '../components/layout/Container'
-import { Content } from '../components/layout/Content'
+import { CardMenu } from '../modules/CardMenu'
+import { Container } from '../modules/layout/Container'
+import { Content } from '../modules/layout/Content'
 import { groupOrder } from '../constants/groupOrder'
 import { sectionOrder } from '../constants/sectionOrder'
