+++ src/pages/dataElement/DataElementList.js
@@ -1,12 +1,12 @@
 import React from 'react'
 
-import { Container } from '../../components/layout/Container'
-import { Content as ContentLayout } from '../../components/layout/Content'
-import { SideBar as SideBarLayout } from '../../components/layout/SideBar'
+import { Container } from '../../modules/layout/Container'
+import { Content as ContentLayout } from '../../modules/layout/Content'
+import { SideBar as SideBarLayout } from '../../modules/layout/SideBar'
 
-import { Content } from '../../components/Content'
-import { PageHeadline } from '../../components/PageHeadline'
-import { Sidebar } from '../../components/Sidebar'
+import { Content } from '../../modules/Content'
+import { PageHeadline } from '../../modules/PageHeadline'
+import { Sidebar } from '../../modules/Sidebar'
 import { dataElementSections } from '../../constants/sections'
 import { groups } from '../../constants/groups'
 import { sectionOrder } from '../../constants/sectionOrder'
